+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Compile GWT modules as part of build process to ensure that the `.gwt.xml` file is consistent.
 * Split gwt modules into `WebPoller.gwt.xml` and `Event.gwt.xml` so that downstream users only
   need to compile the events library if it is used.
+* Migrate `AbstractHttpRequestFactory` and `HttpRequestFactory` classes into the package
+  `org.realityforge.gwt.webpoller.client.http` and create a new gwt module so that downstream users
+  only need to compile gwts http library if it is actually used.
 
 ## 0.9.7: